+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## In Progress ##
 
+- **Fixed:** dictionary.find_leaf_value. Infinite recursion for args: *key="a", dictionary={"a": "b", "b": "a"}*
+- **Added:** New tests. Better coverage
+
 ## v2.1.0 (2022-12-28) ##
 
 - **Added:** Apply the *black* code formatter.